app/requirements.txt
Django>=3.1.3,<3.2.0
djangorestframework>=3.12.2,<3.13.0
gunicorn>=20.0.4,<20.1.0
drf-yasg>=1.20.0,<1.21.0
flake8>=3.8.4,<3.9.0
psycopg2>=2.8.6,<2.9.0
djoser>=2.1.0,<2.2.0
djangorestframework-simplejwt>=4.6.0
django-cors-headers>=3.5.0,<3.6.0
django-debug-toolbar>=3.2,<3.3
django-redis>=4.12.1,<4.13.0
razorpay==1.2.0